[refactor/sbc] 프로젝트 다른 Organizations 으로 이동하기
fsd/pages 레이어를 활용하도록 수정
