+++ README.md
@@ -2,6 +2,14 @@ This is a [Next.js](https://nextjs.org) project bootstrapped with [`create-next-
 ## Next Portfolio
 
 ## 현재 작업중입니다.
+- 해당 포트폴리오를 다른 Organizations 으로 이동했습니다.
+- 포트폴리오 프로젝트 코드를 확인하시려면 아래의 링크를 눌러주세요.
+- <a href="https://github.com/SHIN96BC-Portfolio">포르폴리오 프로젝트 확인하기</a>
+
+## 프로젝트 관리
+- <a href="https://shin96bc.atlassian.net?continue=https%3A%2F%2Fshin96bc.atlassian.net%2Fwelcome%2Fsoftware&atlOrigin=eyJpIjoiMTI2MzE2ZDk0MTlhNDQ3MGE1ODViODMwNWZkNzIxZWEiLCJwIjoiaiJ9">Jira 확인하기</a>
+  
+- <a href="">Figma 확인하기(아직 추가전)</a>
 
 ## 프로젝트 설명
 - Next.js 웹 프론트 코드를 보여드리기 위한 프로젝트 입니다.
@@ -13,7 +21,7 @@ This is a [Next.js](https://nextjs.org) project bootstrapped with [`create-next-
         <img src="https://img.shields.io/badge/npm(v11.7.0)-CB3837?style=flat&logo=npm&logoColor=white"/>
     </div>
     <div>
-        <img src="https://img.shields.io/badge/Next(v15.5.6)-000000?style=flat&logo=next.js&logoColor=white"/>
+        <img src="https://img.shields.io/badge/Next(v16.0.10)-000000?style=flat&logo=next.js&logoColor=white"/>
         <img src="https://img.shields.io/badge/React(v19)-61DAFB?style=flat&logo=react&logoColor=white"/>
         <img src="https://img.shields.io/badge/Typescript-3178C6?style=flat&logo=typescript&logoColor=white"/>
     </div>
