+++ src/app/globals.css
@@ -4,45 +4,70 @@
 @custom-variant dark (&:where(.dark, .dark *));
 
 @theme inline {
-  /* bg-background / text-foreground 같은 유틸이 생기고,
-     실제 CSS엔 var(--bg)/var(--fg)로 출력되어 .dark에서 값이 바뀌면 즉시 반영됨 */
   --color-background: var(--bg);
   --color-foreground: var(--fg);
   --color-primary: var(--primary);
+  --color-primary-hover: var(--primary-hover);
+  --color-secondary: var(--secondary);
+  --color-secondary-hover: var(--secondary-hover);
   --color-muted: var(--muted);
+  --color-muted-foreground: var(--muted-fg);
+  --color-card: var(--card);
+  --color-card-foreground: var(--card-fg);
   --color-border: var(--border);
+  --color-accent: var(--accent);
+  --color-accent-foreground: var(--accent-fg);
+  --color-success: var(--success);
+  --color-warning: var(--warning);
+  --color-error: var(--error);
 }
 
-@theme static {
-  --color-primary-light: #c9c4f9;
-  --color-primary: #9e95f5;
-  --color-primary-dark: #7367f0;
-  --color-secondary-light: #69c1ff;
-  --color-secondary: #36abff;
-  --color-secondary-dark: #0396ff;
-}
-
-/* 디자인 토큰(런타임 변경 가능한 CSS 변수) */
+/* Light */
 :root {
   --bg: #ffffff;
-  --fg: #1a1a1a;
-  --muted: #f5f5f5;
-  --primary: #9e95f5;
-  --border: #e5e5e5;
+  --fg: #111827;
+  --primary: #2563eb;
+  --primary-hover: #1d4ed8;
+  --secondary: #6b7280;
+  --secondary-hover: #4b5563;
+  --muted: #f3f4f6;
+  --muted-fg: #6b7280;
+  --card: #ffffff;
+  --card-fg: #111827;
+  --border: #e5e7eb;
+  --accent: #f3f4f6;
+  --accent-fg: #111827;
+  --success: #16a34a;
+  --warning: #d97706;
+  --error: #dc2626;
 }
 
+/* Dark */
 :root.dark,
 [data-theme="dark"] {
-  --bg: #121212;
-  --fg: #f5f5f5;
-  --muted: #2b2b2b;
-  --primary: #7367f0;
-  --border: #333333;
+  --bg: #0f172a;
+  --fg: #f1f5f9;
+  --primary: #3b82f6;
+  --primary-hover: #60a5fa;
+  --secondary: #94a3b8;
+  --secondary-hover: #cbd5e1;
+  --muted: #1e293b;
+  --muted-fg: #94a3b8;
+  --card: #1e293b;
+  --card-fg: #f1f5f9;
+  --border: #334155;
+  --accent: #1e293b;
+  --accent-fg: #f1f5f9;
+  --success: #22c55e;
+  --warning: #f59e0b;
+  --error: #ef4444;
 }
 
 /* 사용자 에이전트에도 힌트 */
 html {
   color-scheme: light dark;
+  margin: 0;
+  padding: 0;
 }
 
 body {
