+++ src/fsd/entities/site/api/SiteService.ts
@@ -1,6 +1,8 @@
 import { GnbGetRes } from '@FsdEntities/site/model/server';
+import { PortfolioGetRes } from '@FsdEntities/site/model/server/portfolio';
 import { CommonRes } from '@Libs/service-container';
 
 export default interface SiteService {
   getGnb(): Promise<CommonRes<GnbGetRes[]>>;
+  getPortfolio(): Promise<CommonRes<PortfolioGetRes[]>>;
 }
