+++ src/fsd/entities/site/api/SiteServiceImpl.ts
@@ -1,5 +1,6 @@
 import SiteService from '@FsdEntities/site/api/SiteService';
 import { GnbGetRes } from '@FsdEntities/site/model/server';
+import { PortfolioGetRes } from '@FsdEntities/site/model/server/portfolio';
 import { CommonRes, CommonServiceBase } from '@Libs/service-container';
 
 class SiteServiceImpl implements SiteService {
@@ -24,6 +25,14 @@ class SiteServiceImpl implements SiteService {
    */
   getGnb(): Promise<CommonRes<GnbGetRes[]>> {
     return this.base.http.get<CommonRes<GnbGetRes[]>>('/site/gnb');
+  }
+
+  /**
+   * 포트폴리오 컨텐츠 조회
+   * @returns {Promise<CommonRes<PortfolioGetRes>>}
+   */
+  getPortfolio(): Promise<CommonRes<PortfolioGetRes[]>> {
+    return this.base.http.get<CommonRes<PortfolioGetRes[]>>('/site/portfolio');
   }
 }
