+++ src/fsd/pages/default.tsx
@@ -1,0 +1,3 @@
+export default function DefaultPage() {
+  return null;
+}
