+++ src/fsd/pages/interception-modal/login/index.tsx
@@ -1,0 +1,10 @@
+import { LoginForm } from '@FsdFeatures/login/ui';
+import { InterceptionModal } from '@FsdShared/modal/ui';
+
+export default function InterceptionModalLoginPage() {
+  return (
+    <InterceptionModal maxWidth={500} maxHeight={500}>
+      <LoginForm isInterceptionPage />
+    </InterceptionModal>
+  );
+}
