+++ src/fsd/pages/login/index.tsx
@@ -1,0 +1,10 @@
+import { LoginForm } from '@FsdFeatures/login/ui';
+
+export default function LoginPage() {
+  return (
+    <div>
+      <div>Login Page</div>
+      <LoginForm />
+    </div>
+  );
+}
