+++ src/fsd/pages/not-found.tsx
@@ -1,0 +1,3 @@
+export default function RootNotFound() {
+  return <div>404 Not Found</div>;
+}
